% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bendy.R
\name{bendy}
\alias{bendy}
\title{BENDY}
\usage{
bendy(data = HAZ, time = 0:15, hist.lgth = 7)
}
\arguments{
\item{data}{A matrix of values with each row representing an individual and each column representing measurements at each time point (e.g. column 1 is time point 1, etc.). Measurements are assumed to be taken at the same time intervals for every participant.}

\item{time}{A vector of times. This equal the number of columts in the matrix provided (e.g. 0 through 15 months for the example dataset).}

\item{hist.lngth}{The length of known history for the observed process. We use leave one-curve out cross validation for prediction (hist.lgth is 7 for example dataset).}
}
\value{
A matrix of predictions with rows representing each individual and columts representing predictions for each time point.
}
\description{
Dynamic predictions with BENDY are obtained for each subject at each time point t* + j. For all subjects
except subject i we consider data Y[-i,1] and Y[-i,t*], corresponding to first and last HAZ data from the
known HAZ history, as predictive data for BENDY. We use all except the i-th subject to obtain the BENDY
model fit, because we perform the leave one-curve out cross validation for prediction. The BENDY model fit
is done n times to account for dynamic prediction for all n subjects.
The lm function in R is used to fit the BENDY model. For each dynamic prediction, a BENDY model fit is
used. Model fit and prediction are needed for each j for dynamic prediction at times t* + j .
}
\references{
Ivanescu AE, Crainiceanu CM, Checkley W. Dynamic child growth prediction: A comparative methods approach. Statistical Modelling. 2017 Dec;17(6):468-93.
}
