% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{DLM}
\usage{
dlm(data = HAZ, time = 0:15, hist.lgth = 7)
}
\arguments{
\item{data}{A matrix of values with each row representing an individual and each column representing measurements at each time point (e.g. column 1 is time point 1, etc.). Measurements are assumed to be taken at the same time intervals for every participant.}

\item{time}{A vector of times. This equal the number of columts in the matrix provided (0 through 15 months for the example dataset).}

\item{hist.lngth}{The length of known history for the observed process. We use leave one-curve out cross validation for prediction (hist.lgth is 7 for the example dataset).}
}
\value{
A matrix of predictions with rows representing each individual and columts representing predictions for each time point.
}
\description{
DLM uses more information than BENDY. While BENDY uses Y[i,1] and Y[i,t*] for dynamic prediction,
DLM includes data Y[i,1:t*] which contains all the known history before time t* for a subject i. The DLM
model fit is done at each point t* + j.
}
\references{
Ivanescu AE, Crainiceanu CM, Checkley W. Dynamic child growth prediction: A comparative methods approach. Statistical Modelling. 2017 Dec;17(6):468-93.
}
