% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpfr_gam.R
\name{dpfr_gam}
\alias{dpfr_gam}
\title{DPFR using gam}
\usage{
dpfr_gam(data = HAZ, time = 0:15, hist.lngth = 7)
}
\arguments{
\item{data}{A matrix of values with each row representing an individual and each column representing measurements at each time point (e.g. column 1 is time point 1, etc.). Measurements are assumed to be taken at the same time intervals for every participant.}

\item{time}{A vector of times. This equal the number of columts in the matrix provided (e.g. 0 through 15 months for the example dataset).}

\item{hist.lngth}{The length of known history for the observed process. We use leave one-curve out cross validation for prediction (hist.lgth is 7 for the example dataset).}
}
\value{
A matrix of predictions with rows representing each individual and columts representing predictions for each time point.
}
\description{
DPFR uses penalized functional regression to incorporate all the history of Y up to time t* for subject i in
a scalar-on-function regression. At each t* + j the response for subject i is the scalar Y[i,t*+j] and the
functional predictor data consists of Y[i,1:t*]. We show how to obtain DPFR dynamic prediction using
the function pfr from the refund R package and the gam function from the mgcv R package.
}
\details{
Another option for fitting DPFR instead of using pfr, is using the gam function from the mgcv R package. To use gam, there is a
step that involves arranging the data in some specific form prior to calling the gam function for DPFR model
fitting.
}
\references{
Ivanescu AE, Crainiceanu CM, Checkley W. Dynamic child growth prediction: A comparative methods approach. Statistical Modelling. 2017 Dec;17(6):468-93.
}
