% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{MSE}
\usage{
mse(observed = "observed", id.var = "id", predicted = "pred",
  time.var = "time", data)
}
\arguments{
\item{observed}{observed growth values (e.g. height or weight)}

\item{id.var}{variable that identifies individual subjects}

\item{predicted}{predicted values from models fitted to observed data}
}
\value{
None
}
\description{
Function for calculating subject-specific mean squared error
}
