% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmse.R
\name{nmse}
\alias{nmse}
\title{nMSE}
\usage{
nmse(observed = "observed", predicted = "pred", id.var = "id", data)
}
\arguments{
\item{observed}{observed growth values (e.g. height or weight)}

\item{predicted}{predicted values from models fitted to observed data}

\item{id.var}{variable that identifies individual subjects}
}
\value{
data.frame with id and subject-specific location-normalized MSE estimates
}
\description{
Function for calculating subject-specific location-normalized mean squared error
}
\references{
Grigsby MR, Di J, Leroux A, Zipunnikov V, Xiao L, Crainiceanu C, Checkley W. Novel metrics for growth model selection. Emerging themes in epidemiology. 2018 Dec;15(1):4.
}
