% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmse.R
\name{wmse}
\alias{wmse}
\title{wMSE}
\usage{
wmse(id.var = "id", time.var = "time", weight.var = "weights",
  type = "mse", data)
}
\arguments{
\item{id.var}{variable that identifies individual subjects}

\item{time.var}{time variable (e.g. age) used when calculating nMSE and wMSE}

\item{weight.var}{Variable used to weight MSE or nMSE. This should be a vector of values that will be used to divide subject-specific MSE estimates by. An example could be using subject-specific growth trajectories (i.e. weighting individuals with slowest growth).}

\item{observed}{observed growth values (e.g. height or weight)}

\item{predicted}{predicted values from models fitted to observed data}
}
\value{
None
}
\description{
Function for weighting MSE, nMSE, or aMSE by the specified variable for weights.
}
