% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amse.R
\name{amse}
\alias{amse}
\title{aMSE}
\usage{
amse(observed = "observed", predicted = "pred", id.var = "id",
  time.var = "time", breaks = c(0, 6, 12, 18, 24), data)
}
\arguments{
\item{observed}{observed growth values (e.g. height or weight)}

\item{predicted}{predicted values from models fitted to observed data}

\item{id.var}{variable that identifies individual subjects}

\item{time.var}{time variable (e.g. age) used when calculating nMSE and wMSE}

\item{breaks}{vector indicating location of age cutoffs for calculating MSE per group}
}
\value{
None
}
\description{
Function for calculating age-stratified subject-specific mean squared error
}
