% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{MSE}
\usage{
mse(observed = "height", id.var = "id", predicted = "pred",
  time.var = "time", data)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Function for calculating subject-specific mean squared error
}
