% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmse.R
\name{nmse}
\alias{nmse}
\title{nMSE}
\usage{
nmse(observed = "height", id.var = "id", predicted = "pred",
  time.var = "time", data)
}
\arguments{
\item{observed}{observed growth values (e.g. height or weight)}

\item{id.var}{variable that identifies individual subjects}

\item{predicted}{predicted values from models fitted to observed data}

\item{time.var}{time variable (e.g. age) used when calculating nMSE and wMSE}
}
\value{
None
}
\description{
Function for calculating subject-specific location-normalized mean squared error
}
