% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmse.R
\name{wmse}
\alias{wmse}
\title{wMSE}
\usage{
wmse(id.var = "id", time.var = "time", weight.var = "weights",
  type = "mse", data)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Function for weighting MSE, nMSE, or aMSE by the specified variable for weights.
}
